Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 // Your code here...
 #include<stdio.h>
 int main(){
-    int a ,area;
+    float a ,area;
     scanf("%.2f",&a);
     area=3.14 * a * a;
     printf("Area: %.2f",area);
